+++ tests/test_snapshots.py
@@ -394,9 +394,7 @@ class TestVariables:
             # The params typed below should be dimmed since they dont resolve
             await pilot.press(*"$nope/${nope}")
 
-        assert snap_compare(
-            POSTING_MAIN, run_before=run_before, terminal_size=(118, 34)
-        )
+        assert snap_compare(POSTING_MAIN, run_before=run_before)
 
     def test_resolved_variables_highlight_and_preview(self, snap_compare):
         """Check that the resolved variables are highlighted in the URL